Pequeña refactorizacion de tipos para consistencia y eliminacion de constructores en servicios(se usa autowired)

+++ planifyAPI/src/main/java/com/chilltime/planifyapi/repository/CalendarioRepository.java
@@ -4,7 +4,7 @@ import com.chilltime.planifyapi.entity.Calendario;
 import com.chilltime.planifyapi.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CalendarioRepository extends JpaRepository<Calendario, Integer> {
+public interface CalendarioRepository extends JpaRepository<Calendario, Long> {
 
 
     boolean existsByNombreAndUsuario(String nombre, Usuario usuario);
